# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Simple Calculator RestAPI",
        "contact": {}
    },
    "paths": {
        "/add": {
            "get": {
                "description": "plus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "add",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "used for calc",
                        "name": "num1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "used for calc",
                        "name": "num2",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "answer",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/subtract": {
            "get": {
                "description": "plus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "subtract",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "used for calc",
                        "name": "num1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "used for calc",
                        "name": "num2",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "answer",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}